+++ src/api/clients/ollama.ts
@@ -2,7 +2,7 @@ import { Notice } from 'obsidian';
 import OpenAI from 'openai';
 import Markpilot from 'src/main';
 import { validateURL } from 'src/utils';
-import { APIClient } from '../client';
+import { APIClient } from '..';
 import { CostsTracker } from '../costs';
 import { OpenAICompatibleAPIClient } from './openai-compatible';
