+++ src/index.d.ts
@@ -2,3 +2,13 @@ declare module '*.svg' {
   const content: string;
   export default content;
 }
+
+declare module '*.md' {
+  const content: string;
+  export default content;
+}
+
+declare module '*.txt' {
+  const content: string;
+  export default content;
+}
